+++ tests/testthat/test-generate_pt_nowcast_mat.R
@@ -90,15 +90,14 @@ test_that("generate_pt_nowcast_mat errors when n is incompatible", {
       test_triangle,
       n = 8
     )
-  ) # nolint
+  )
   # Custom n_history_delay is too low
   expect_error(
     generate_pt_nowcast_mat(
       test_triangle,
       n = 3
-    ),
-    regexp = "The number of observations specified for delay estimation is less"
-  ) # nolint
+    )
+  )
 })
 
 test_that("generate_pt_nowcast_mat can take in another delay PMF", {
@@ -110,4 +109,52 @@ test_that("generate_pt_nowcast_mat can take in another delay PMF", {
   last_row <- result[5, ]
   pmf <- last_row / sum(last_row)
   expect_equal(pmf, external_delay_pmf, tolerance = 0.01)
+})
+
+test_that("generate_pt_nowcast_mat generates the correct result on a ragged triangle", { # nolint
+  delay_pmf <- c(0.4, 0.3, 0.2, 0.05, 0.05)
+  partial_counts <- c(80, 100, 180, 80, 140)
+
+  # Create a complete triangle based on the known delay PMF
+  triangle <- lapply(partial_counts, function(x) x * delay_pmf)
+  triangle <- do.call(rbind, triangle)
+  triangle <- generate_triangle(triangle, structure = c(1, 2))
+
+  result <- generate_pt_nowcast_mat(
+    triangle
+  )
+  cols <- colSums(result[3:5, ])
+  pmf <- cols / sum(cols)
+  expect_equal(pmf, delay_pmf, tolerance = 0.01)
+})
+
+test_that("generate_pt_nowcast_mat generates the correct result on a ragged triangle with fewer rows than columns", { # nolint
+  delay_pmf <- c(0.4, 0.3, 0.2, 0.05, 0.05)
+  partial_counts <- c(80, 100, 180)
+
+  # Create a complete triangle based on the known delay PMF
+  triangle <- lapply(partial_counts, function(x) x * delay_pmf)
+  triangle <- do.call(rbind, triangle)
+  triangle <- generate_triangle(triangle, structure = c(1, 2))
+
+  result <- generate_pt_nowcast_mat(
+    triangle
+  )
+  cols <- colSums(result[2:3, ])
+  pmf <- cols / sum(cols)
+  expect_equal(pmf, delay_pmf, tolerance = 0.01)
+})
+
+test_that("generate_pt_nowcast_mat errors when there are insufficient observations", { # nolint
+  delay_pmf <- c(0.4, 0.3, 0.2, 0.05, 0.05)
+  partial_counts <- c(80, 100)
+
+  # Create a complete triangle based on the known delay PMF
+  triangle <- lapply(partial_counts, function(x) x * delay_pmf)
+  triangle <- do.call(rbind, triangle)
+  triangle <- generate_triangle(triangle, structure = c(1, 2))
+
+  expect_error(generate_pt_nowcast_mat(
+    triangle
+  ))
 })